// frontend/src/app/api/auth/model/authResponseDto.ts
/**
 * PaperNote Auth API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';


/**
 * Response model containing JWT access token, refresh token, expiration information and user details
 */
export interface AuthResponseDto { 
    accessToken?: string | null;
    refreshToken?: string | null;
    expiresAt?: string;
    user?: UserDto;
}