// frontend/src/app/api/auth/model/loginDto.ts
/**
 * PaperNote Auth API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request model for user authentication with username and password credentials
 */
export interface LoginDto { 
    username?: string | null;
    password?: string | null;
}