// frontend/src/app/api/auth/model/registerDto.ts
/**
 * PaperNote Auth API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request model for registering a new user account with username and password
 */
export interface RegisterDto { 
    username?: string | null;
    password?: string | null;
}