// frontend/src/app/api/auth/model/refreshTokenDto.ts
/**
 * PaperNote Auth API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request model containing a refresh token for obtaining new JWT access and refresh tokens
 */
export interface RefreshTokenDto { 
    refreshToken?: string | null;
}